Parser improved (still lot of room for improvement - rewrite needed at some point). Clippy alter will be fixed in the main branch.

+++ pobsdlib/src/models/item.rs
@@ -13,7 +13,7 @@ pub struct Item {
     /// name of the item.
     pub name: String,
     /// vector of ids of the games associated to the items.
-    pub games: Vec<usize>,
+    pub games: Vec<u64>,
 }
 
 impl fmt::Display for Item {
